minor improvements
the new csx64 standard enables assembly script execution.
added a shebang to all test files to make use of this.
this means tests no longer require assembling or linking.
additionally the shell scripts to make this more convenient are no longer needed - removed.

the old malloc implementation would use sys_brk(0) to get the malloc field starting location if it wasn't already initialized.
this could theoretically be wasteful if for some reason the current break point isn't directly at the end of the stack segment (like it's initialized to).
renamed malloc_beg/end to __malloc_beg/end and made them global (with a note to never modify them).
these are initialized by _start to rbp, which the csx64 standard initializes to the address of the end of the stack segment.
this also means the malloc logic for handling uninitialized beg/end pointers is unnecessary, but is left in for redundancy safety - since the branch will likely never be taken (unless using a custom _start file) it hardly impacts performance.