added file reading functions
added fgetc, ungetc.
also added fpeek as a (non-standard) utility function.

added tests in files.asm for making sure the file processing works as desired.
merged the fsmanip.asm stuff in stdio.asm (was in a different one because i wasn't ready to impl the rest of the stdio header at the time).