added arglist.asm
added arglist.asm to the stdlib folder.
this includes various functions for iterating through all arguments passed to a csxdecl function.
this can be used by any csxdecl function, but is primarily intended for creating varargs functions.

added arglist.asm to the test folder to do a simple test making sure it works as intended for a varargs function.