print fixes
there was a running bug for a while where puts() wasn't also putting a new line char after string.
i held off on fixing it for quite some time because i think it's stupid that it does that.
but now that it's C standard compliant, had to switch over all the puts() uses with fputs() or remove the new line in the string i was printing (one or the other).

added the -t flag to the test/fpu.asm shebang since i always run it with that anyway.