printf update: hex, octal, and sizes
added hex and octal format options for printf
these, along with decimal, are all handled by __printf_u64_raw (renamed from __printf_u64_decimal_raw).
this function takes the radix and alpha base.
this general function was used to implement the octal and hex (lowercase and uppercase).

all of the logic to decide this in the actual printf utility function has been rolled into using an array of function pointers.
the format pack size field is used as the index into this array.
there are separate arrays for arg fetching and printing to allow for more control over everything for later updates (e.g. floating point).