added %f printf format flag
added the first (limited) implementation of the %f format specifier for printf.
currently if no precision is specified it is defaulted to 6.
not tested for anything other than using all defaults.
handles positive and negative normal and denormal values.
supports infinity/nan but strips it of the sign information (will be addressed better in later updates).
only tested for small values.
also performs rounding on the stringified representation, but is only tested for a handfull of values.

changed the integer fetchers to all be 64-bit.
the standard calling conventions for csxdecl varargs functions from here on out will require all args be totally extended.
for signed/unsigned integers, this means u64 or i64.
for floating point types this means f64.
the arglist fetchers for other sizes will still be supported, but will no longer be used by the standard library functions.

the %f flag for prntf accepts both %f and %lf, but any other size prefix is invalid.

added test cases for movd/movq for cpu/vpu register transfers.