added snprintf/vsnprintf
added the snprintf and vsnprintf functions.
they're very similar to the unbounded string formatting functions.
the only difference is the extra logic in the sprinter for no-oping buffer overflows.

after doing a bit more research, standard cdecl appears to push args in 8-byte chunks, with padding for any missing bytes.
and for varargs parameters, they're actually extended (not just padded).
e.g. passing a float as normal param is actual size, but padded if put on stack.
but passing a foat as a varargs param is extended to double both for register and stack params.
thus i've decided to remove all the non-fully-extended arglist_T accessor functions.
only 64-bit forms are available now.
additionally they are renamed to arglist_i and arglist_f instead of ending with 64.
this is done to emphasize that all int/float parameters are the same type (after being extended).
for example, if you need a normal (non varargs) float param, use arglist_f and assume you got back a 32-bit floating point value.

added assertion that vsprintf returns the correct value.
this of course applies to all variants of printf because they share a base implementation.
this caught a bug where we were computing the total number of chars printed correctly but forgot to move it into the return value at the end.